move all base functionality back to DriveObject
