+++ pygsuite/drive/__init__.py
@@ -1,5 +1,6 @@
 from .drive import Drive, UserType, PermissionType
+from .drive_object import DriveObject
 from .file import File
 from .folder import Folder
 
-__all__ = ["Drive", "File", "Folder", "UserType", "PermissionType"]
+__all__ = ["Drive", "DriveObject", "File", "Folder", "UserType", "PermissionType"]
